refs #1071  * 체크업시 Memcached를 flushing하도록 함.  * 세션 쓰기 과정에서 객체가 유실되는 문제. (어느 시점에서 unset되는지 찾아야 한다)

+++ interface/blog/checkup.php
@@ -44,7 +44,7 @@ function showCheckupMessage($stat = true) {
 }
 
 function clearCache() {
-	global $database, $changed, $errorlog;
+	global $database, $changed, $errorlog, $memcache;
 	static $isCleared = false;
 	if($isCleared == true) return true;
 	if($blogids = POD::queryColumn("SELECT blogid FROM {$database['prefix']}PageCacheLog")) {
@@ -57,7 +57,11 @@ function clearCache() {
 		if($errorlog == false) echo '<span class="result success">', _text('성공'), '</span></li>';
 		else echo '<span class="result fail">', _text('실패'), '</span></li>';
 	}
-
+	if(!is_null($memcache)) {
+		echo '<li>', _textf('Memcached 캐시를 초기화합니다.'), ': ';
+		if($memcache->flush())  echo '<span class="result success">', _text('성공'), '</span></li>';
+		else echo '<span class="result fail">', _text('실패'), '</span></li>';
+	}
 	echo '<li>', _textf('공지사항 캐시를 초기화합니다.'), ': ';
 	if(POD::execute("DELETE FROM {$database['prefix']}ServiceSettings WHERE name like 'TextcubeNotice%'"))
 		echo '<span class="result success">', _text('성공'), '</span></li>';
